+++ src/components/Polystat.tsx
@@ -242,6 +242,7 @@ export const Polystat: React.FC<PolystatOptions> = (options) => {
 
   if (options.globalAutoScaleFonts) {
     const result = autoFontScaler(
+      options.globalTextFontFamily,
       textAreaWidth,
       textAreaHeight,
       options.globalShowValueEnabled,
@@ -415,6 +416,7 @@ export const Polystat: React.FC<PolystatOptions> = (options) => {
                     secondarySortDirection={options.tooltipSecondarySortDirection}
                     displayMode={options.tooltipDisplayMode}
                     tooltipDisplayTextTriggeredEmpty={options.tooltipDisplayTextTriggeredEmpty}
+                    tooltipFontFamily={options.globalTooltipsFontFamily}
                   />
                 )}
                 <text
@@ -422,7 +424,7 @@ export const Polystat: React.FC<PolystatOptions> = (options) => {
                   x={coords.x + alignments.labelTextAlignmentX}
                   y={coords.y + alignments.labelWithValueTextAlignment}
                   textAnchor="middle"
-                  fontFamily="Roboto"
+                  fontFamily={options.globalTextFontFamily}
                   fontSize={activeLabelFontSize + 'px'}
                   style={{
                     fill: options.globalTextFontAutoColorEnabled
@@ -449,7 +451,7 @@ export const Polystat: React.FC<PolystatOptions> = (options) => {
                   x={coords.x + alignments.labelValueAlignmentX}
                   y={coords.y + alignments.valueWithLabelTextAlignment}
                   textAnchor="middle"
-                  fontFamily="Roboto"
+                  fontFamily={options.globalTextFontFamily}
                   fontSize={activeValueFontSize + 'px'}
                   style={{
                     fill: options.globalTextFontAutoColorEnabled
@@ -569,6 +571,7 @@ const getAlignments = (
 };
 
 const autoFontScaler = (
+  fontFamily: string,
   textAreaWidth: number,
   textAreaHeight: number,
   valueEnabled: boolean,
@@ -613,6 +616,7 @@ const autoFontScaler = (
   // if it is too small, hide everything
   let activeLabelFontSize = computeTextFontSize(
     maxLabel,
+    fontFamily,
     minFont,
     maxFont,
     maxLinesToDisplay,
@@ -621,6 +625,7 @@ const autoFontScaler = (
   );
   let activeValueFontSize = computeTextFontSize(
     maxValue,
+    fontFamily,
     minFont,
     maxFont,
     maxLinesToDisplay,
@@ -634,6 +639,7 @@ const autoFontScaler = (
     maxLabel = maxLabel.substring(0, numOfChars + 2);
     activeLabelFontSize = computeTextFontSize(
       maxLabel,
+      fontFamily,
       minFont,
       maxFont,
       maxLinesToDisplay,
@@ -645,6 +651,7 @@ const autoFontScaler = (
       maxLabel = maxLabel.substring(0, numOfChars + 2);
       activeLabelFontSize = computeTextFontSize(
         maxLabel,
+        fontFamily,
         minFont,
         maxFont,
         maxLinesToDisplay,
@@ -656,6 +663,7 @@ const autoFontScaler = (
         maxLabel = maxLabel.substring(0, numOfChars + 2);
         activeLabelFontSize = computeTextFontSize(
           maxLabel,
+          fontFamily,
           minFont,
           maxFont,
           maxLinesToDisplay,
@@ -678,6 +686,7 @@ const autoFontScaler = (
 
 const computeTextFontSize = (
   text: string,
+  font: string,
   minFont: number,
   maxFont: number,
   linesToDisplay: number,
@@ -686,7 +695,7 @@ const computeTextFontSize = (
 ): number => {
   return getTextSizeForWidthAndHeight(
     text,
-    '?px sans-serif', // use sans-serif for sizing
+    `?px ${font}`,
     textAreaWidth,
     textAreaHeight / linesToDisplay, // multiple lines of text
     minFont,