+++ src/components/tooltips/Tooltip.tsx
@@ -22,6 +22,7 @@ export interface TooltipProps {
   secondarySortDirection: number;
   displayMode: string;
   tooltipDisplayTextTriggeredEmpty: string;
+  tooltipFontFamily: string;
 }
 
 export const Tooltip = ({
@@ -39,6 +40,7 @@ export const Tooltip = ({
   secondarySortDirection,
   displayMode,
   tooltipDisplayTextTriggeredEmpty,
+  tooltipFontFamily,
 }: TooltipProps) => {
   const styles = useStyles2(getTooltipStyles);
 
@@ -146,7 +148,7 @@ export const Tooltip = ({
 
   const generateContent = () => {
     return (
-      <table className={styles.tooltipTable}>
+      <table style={{ fontFamily: tooltipFontFamily }} className={styles.tooltipTable}>
         <thead>
           {data && data.isComposite && getCompositeHeader(data)}
           <tr>